+++ CHANGES.rst
@@ -11,4 +11,5 @@ Unreleased
 * Added :class:`IndexedDict`
 * Improve efficiency for large bag operations
 * Add :meth:`setlist.swap`
-* Add :meth:`bag.count` and :class:`CountsView`
+* Add :meth:`bag.count`, :class:`CountsView` & :class:`UniqueElementsView`
+* Add :meth:`bag.is_subset` and :meth:`is_superset`